feat(auth): implement LoginForm component with improved UI and form handling
- Introduced a new LoginForm component that utilizes react-hook-form for form management.
- Enhanced user experience with a structured form layout using Form, FormControl, FormField, and Button components.
- Integrated Spinner for loading state during submission.
- Reset form fields upon successful submission.
